Moved MetricType to entities
